% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{idfetcher_clear_credentials}
\alias{idfetcher_clear_credentials}
\title{Clear Zotero credentials}
\usage{
idfetcher_clear_credentials()
}
\value{
Invisibly returns TRUE.
}
\description{
Removes Zotero credentials from the current R session.
}
