% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{idfetcher_get_credentials}
\alias{idfetcher_get_credentials}
\title{Get Zotero credentials}
\usage{
idfetcher_get_credentials()
}
\value{
A list containing user_id and api_key.
}
\description{
Retrieves the Zotero credentials currently stored for the R session.
}
